+++ README.md
@@ -1,15 +1,28 @@
+![Image](https://moviemultiverse.github.io/core/icoon.png)
+
+
 # The Core 
-A platform server to store movies at mega large scale
+A Video Content Provider for Blackhole
 
 ## Description
-A platform server to store movies with postgres (graphql) loaded
-Uses JS , Dart , Python3 as main Building blocks
+A platform service engineered for the storage of video content integrates PostgreSQL with GraphQL for data management. It utilizes JavaScript, Dart, and Python 3 as its core programming languages. For enhanced performance and scalability, the architecture incorporates both Redis and MongoDB databases.
 
 ## How to Contribute 
-`
 
-`
+> fork the repos
 
 ## Documentation
+Have a look at our Documentation
+[https://moviemultiverse.github.io/core](https://moviemultiverse.github.io/core/)
 
-[Docs](https://moviemultiverse.github.io/core/)
+
+## Install
+
+#### Linux || WSL
+
+```bash
+git clone https://github.com/moviemultiverse/core
+cd core
+nano .env -> {add your crendentials}
+npx nodemon index.js
+```
