+++ index.js
@@ -18,10 +18,6 @@ const { Octokit } = require("@octokit/rest");
 const credentials = require('./drive-download-389811-b229f2e27ed8.json');
 const githubToken = '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'; 
 require('dotenv').config();
-const { ApolloServer } = require('@apollo/server');
-const { expressMiddleware } = require('@apollo/server/express4');
-const { GraphQLJSON, GraphQLJSONObject } = require('graphql-type-json');
-const { ApolloServerPluginDrainHttpServer } = require('@apollo/server/plugin/drainHttpServer');
 const http = require('http');
 const cors = require('cors');
 const bodyParser = require('body-parser');
@@ -32,6 +28,8 @@ const { get_telecore_data , search_telecore_data , set_telecore_data} = require(
 const express = require('express');
 const app = express();
 app.use(express.json());
+const port = process.env.PORT || 3000;
+
 
 async function insertuser(authClient, fileId, emailAddress, role) {
   try {
@@ -59,107 +57,39 @@ const uuid_route = require("./routes/uuid_route");
 const public_api = require("./routes/public_api");
 const admin = require("./routes/admin");
 const main_route = require("./routes/main");
-const { typeDefs, resolvers } = require("./controllers/graphql");
-const port = process.env.PORT || 3000;
-const server = new ApolloServer({ typeDefs, resolvers });
 
-server.start().then(() => {
-  app.use(expressMiddleware(server));
-});
 
 app.use("/", uuid_route);
 app.use("/", public_api);
 app.use("/", admin);
 app.use("/", main_route);
-app.listen(port, () => {
-  console.log(` Server ready at http://localhost:${port}`);
+
+
+
+
+
+
+//************************************************APOLLO GRAPHQL SERVER********************************************************* */
+const { ApolloServer } = require('@apollo/server');
+const { expressMiddleware } = require('@apollo/server/express4');
+const { typeDefs, resolvers } = require("./controllers/graphql");
+const server = new ApolloServer({ typeDefs, resolvers });
+server.start().then(() => {
+  app.use(expressMiddleware(server));
 });
-/**************************************************************************************************************************** */
+/******************************************************************************************************************************* */
 
 
 
-
-
-//-----------------------------------FOR TELEGRAM AUTH BOT -------------------------------------------------------
-const { set2redis, get2redis } = require('./redis/redis_data');
-const TelegramBot = require('node-telegram-bot-api');
-const token = process.env.TOKEN ;
-const bot = new TelegramBot(token, { polling: true });
-//https://github.com/yagop/node-telegram-bot-api/issues/406#issuecomment-1270573068
-//https://t.me/blackhole_movie_bot?start=YT0xMjMmYj1nZGZnZC1nZGZnZGZnZGY
-//btoa('a=123&b=gdfgd-gdfgdfgdf')
-//convert the string to base64 param
-bot.on('message', async (msg) => {
-  const admin1 = await get2redis("admin1");
-  const admin2 = await get2redis("admin2");
-  console.log(admin2);
-  console.log(msg);
-  const chatId = msg.chat.id;
-  if(msg.from.id == admin1 && msg.document.file_name != undefined && msg.message_id != undefined){
-    console.log(msg.message_id, msg.document.file_name);
-    const message_id = msg.message_id;
-    const file_name = msg.document.file_name ;
-    if(set_telecore_data([{admin: admin1, message_id: message_id ,file_name: file_name}]))
-    bot.sendMessage(chatId,file_name + " saved successfully");
-    }
-  else if(msg.from.id == admin2 && msg.document.file_name != undefined && msg.message_id != undefined){
-      console.log(msg.message_id, msg.document.file_name);
-      const message_id = msg.message_id;
-      const file_name = msg.document.file_name ;
-      if(set_telecore_data([{admin: admin1, message_id: message_id ,file_name: file_name}]))
-      bot.sendMessage(chatId,file_name + " saved successfully");
-      }
-  else if((msg.from.id != admin1 || msg.from.id != admin2) && msg.text){
+//************************************************TELEGRAM AUTH BOT************************************************************* */
+const { TelecoreBot } = require('./telecore/bot.js');
+(async () => {
   try {
-      const payload = msg.text.substring(6);
-      if (payload.length) {
-        const url = Buffer.from(payload, 'base64').toString();
-        const params = Object.fromEntries(new URLSearchParams(url).entries());
-        // Result: { a: '123', b: 'gdfgd-gdfgdfgdf' }
-        console.log(params);
-          try {
-            console.log(params.text + "success");
-            bot.sendMessage(chatId, 'success');
-            const data1 = await search_telecore_data(params.text);
-            console.log("data",data1);
-            if(data1.admin== admin1 )
-            bot.forwardMessage(chatId,  admin1 , data1.message_id);
-            else
-            bot.forwardMessage(chatId,  admin2 , data1.message_id);
-            //dev1 [message id is from dev1<->bot]
-            //bot.forwardMessage(chatId,  admin2 , data1.id);
-          } catch (error) {
-            console.error('Error fetching telecore data:', error);
-            bot.sendMessage(chatId, 'An error occurred while fetching telecore data');
-          }
-    } else {
-      console.warn('Received message without text:', msg);
-    }
+    await TelecoreBot();
   } catch (error) {
-    console.error('Error in the message handler:', error);
-    bot.sendMessage(msg.chat.id, 'An error occurred while processing your message');
-  }}
-});
-/*
-bot.on('message', async (msg) => {
-  if(msg.from.id == admin1 && msg.document.file_name != undefined && msg.message_id != undefined){
-    console.log(msg.message_id, msg.document.file_name);
-    const message_id = msg.message_id;
-    const file_name = msg.document.file_name ;
-    set_telecore_data([{admin:  admin1 , message_id: message_id ,file_name: file_name}]);
-    }
-  else if(msg.from.id == admin2 && msg.document.file_name != undefined && msg.message_id != undefined){
-      console.log(msg.message_id, msg.document.file_name);
-      const message_id = msg.message_id;
-      const file_name = msg.document.file_name ;
-      set_telecore_data([{admin:  admin2 , message_id: message_id ,file_name: file_name}]);
-      }
-  //bot count message_id on inc++ mode for each user
-});
-*/
-//******************************************************************************************************************************* */
-
-
+    console.error(error);}
+})();
+//***************************************************************************************************************************** */
 
 
 
@@ -357,4 +287,18 @@ else if (xGoogResourceState == 'sync') {
 }
 
   res.json("posted");
+});
+
+
+
+
+
+
+
+
+
+
+
+app.listen(port, () => {
+  console.log(` Server ready at http://localhost:${port}`);
 });
